+++ libs/ui/shared/src/lib/components/HeroSection/index.stories.tsx
