+++ libs/ui/shared/src/lib/components/HeroSection/mocks.ts
