+++ libs/ui/shared/src/lib/components/HeroSection/types.ts
